+++ lib/src/widgets/chatui_textfield.dart
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 import 'dart:async';
-import 'dart:io' show Platform;
+import 'dart:io' show File, Platform;
 
 import 'package:audio_waveforms/audio_waveforms.dart';
 import 'package:chatview/src/utils/constants/constants.dart';
@@ -85,6 +85,9 @@ class _ChatUITextFieldState extends State<ChatUITextField> {
   TextFieldConfiguration? get textFieldConfig =>
       sendMessageConfig?.textFieldConfig;
 
+  CancelRecordConfiguration? get cancelRecordConfiguration =>
+      sendMessageConfig?.cancelRecordConfiguration;
+
   OutlineInputBorder get _outLineBorder => OutlineInputBorder(
         borderSide: const BorderSide(color: Colors.transparent),
         borderRadius: widget.sendMessageConfig?.textFieldConfig?.borderRadius ??
@@ -145,11 +148,16 @@ class _ChatUITextFieldState extends State<ChatUITextField> {
             children: [
               if (isRecordingValue && controller != null && !kIsWeb)
                 AudioWaveforms(
-                  size: Size(MediaQuery.of(context).size.width * 0.75, 50),
+                  size: Size(
+                      MediaQuery.of(context).size.width *
+                          (cancelRecordConfiguration == null ? 0.75 : 0.65),
+                      50),
                   recorderController: controller!,
                   margin: voiceRecordingConfig?.margin,
                   padding: voiceRecordingConfig?.padding ??
-                      const EdgeInsets.symmetric(horizontal: 8),
+                      EdgeInsets.symmetric(
+                        horizontal: cancelRecordConfiguration == null ? 8 : 5,
+                      ),
                   decoration: voiceRecordingConfig?.decoration ??
                       BoxDecoration(
                         color: voiceRecordingConfig?.backgroundColor,
@@ -274,7 +282,19 @@ class _ChatUITextFieldState extends State<ChatUITextField> {
                                   color:
                                       voiceRecordingConfig?.recorderIconColor,
                                 ),
-                          )
+                          ),
+                        if (isRecordingValue &&
+                            cancelRecordConfiguration != null)
+                          IconButton(
+                            onPressed: () {
+                              cancelRecordConfiguration?.onCancel?.call();
+                              _cancelRecording();
+                            },
+                            icon: cancelRecordConfiguration?.icon ??
+                                const Icon(Icons.cancel_outlined),
+                            color: cancelRecordConfiguration?.iconColor ??
+                                voiceRecordingConfig?.recorderIconColor,
+                          ),
                       ],
                     );
                   }
@@ -285,6 +305,27 @@ class _ChatUITextFieldState extends State<ChatUITextField> {
         },
       ),
     );
+  }
+
+  FutureOr<void> _cancelRecording() async {
+    assert(
+      defaultTargetPlatform == TargetPlatform.iOS ||
+          defaultTargetPlatform == TargetPlatform.android,
+      "Voice messages are only supported with android and ios platform",
+    );
+    if (!isRecording.value) return;
+    final path = await controller?.stop();
+    if (path == null) {
+      isRecording.value = false;
+      return;
+    }
+    final file = File(path);
+
+    if (await file.exists()) {
+      await file.delete();
+    }
+
+    isRecording.value = false;
   }
 
   Future<void> _recordOrStop() async {